feat: remove unneeded hooks
