+++ src/components/common/header/ButtonPopover.tsx
@@ -1,7 +1,24 @@
 import * as React from "react";
 import Button from "@mui/material/Button";
 import Popover from "@mui/material/Popover";
-import useAnchorEl from "../../../hooks/useAnchorEl";
+
+interface AnchorElHook {
+  anchorEl: null | HTMLElement;
+  setAnchor: React.MouseEventHandler;
+  clearAnchor: React.MouseEventHandler;
+}
+
+const useAnchorEl = (): AnchorElHook => {
+  const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null);
+
+  const setAnchor = (e: React.MouseEvent<HTMLButtonElement>): void => {
+    setAnchorEl(e.currentTarget);
+  };
+  const clearAnchor = (): void => {
+    setAnchorEl(null);
+  };
+  return { anchorEl, setAnchor, clearAnchor };
+};
 
 interface ButtonPopoverProps {
   children: React.ReactElement;
